+++ rxnetty-contexts/src/test/java/io/reactivex/netty/contexts/BidirectionalTestContext.java
@@ -16,8 +16,6 @@
 
 package io.reactivex.netty.contexts;
 
-import com.netflix.server.context.BiDirectional;
-
 /**
  * @author Nitesh Kant
  */
