+++ rxnetty-contexts/src/test/java/io/reactivex/netty/contexts/BidirectionalTestContextSerializer.java
@@ -16,9 +16,6 @@
 
 package io.reactivex.netty.contexts;
 
-import com.netflix.server.context.ContextSerializationException;
-import com.netflix.server.context.ContextSerializer;
-
 /**
  * @author Nitesh Kant (nkant@netflix.com)
  */
