+++ rxnetty-contexts/src/test/java/io/reactivex/netty/contexts/MergeableBidirectionalTestContextSerializer.java
@@ -16,9 +16,6 @@
 
 package io.reactivex.netty.contexts;
 
-import com.netflix.server.context.ContextSerializationException;
-import com.netflix.server.context.ContextSerializer;
-
 import java.util.regex.Pattern;
 
 /**
